refactor: fix reviewer comms
