+++ src/main/java/ru/practicum/shareit/common/dto/BaseResponse.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.base;
+package ru.practicum.shareit.common.dto;
 
 public record BaseResponse(String message) {
 }
