+++ src/main/java/ru/practicum/shareit/user/UserController.java
@@ -4,7 +4,7 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
-import ru.practicum.shareit.base.BaseResponse;
+import ru.practicum.shareit.common.dto.BaseResponse;
 import ru.practicum.shareit.user.dto.UserDto;
 
 @Slf4j
