+++ src/main/java/ru/practicum/shareit/user/UserService.java
@@ -1,6 +1,6 @@
 package ru.practicum.shareit.user;
 
-import ru.practicum.shareit.base.BaseResponse;
+import ru.practicum.shareit.common.dto.BaseResponse;
 import ru.practicum.shareit.user.dto.UserDto;
 
 public interface UserService {
